mongodb: move the conn-pool drift canary into an integration test file
TestIntegrationConnPoolErrorShapes boots a container and is integration-
gated, so it belongs in a {feature}_integration_test.go per the test file
conventions. It stays in package mongodb (not mongodb_test) because it
probes the unexported isConnPoolError.
