+++ internal/impl/mongodb/common_test.go
@@ -18,18 +18,13 @@ import (
 	"context"
 	"errors"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/testcontainers/testcontainers-go"
-	"github.com/testcontainers/testcontainers-go/wait"
-	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 	"go.mongodb.org/mongo-driver/v2/mongo/options"
 
 	"github.com/redpanda-data/benthos/v4/public/service"
-	"github.com/redpanda-data/benthos/v4/public/service/integration"
 )
 
 func TestIsConnPoolError(t *testing.T) {
@@ -79,98 +74,6 @@ func TestIsConnPoolError(t *testing.T) {
 			assert.Equal(t, test.want, isConnPoolError(test.err))
 		})
 	}
-}
-
-// TestIntegrationConnPoolErrorShapes is a drift canary for isConnPoolError.
-//
-// TestIsConnPoolError above pins the classifier against error strings captured
-// by hand from mongo-driver v2.5.0, which means a driver bump that reworded any
-// of those messages would leave that test passing while the classifier silently
-// stopped recognising real failures. This test feeds isConnPoolError errors
-// produced by the live driver against a real server instead, so the same bump
-// fails CI. The literal substring assertions are deliberate: they name which
-// message shape moved.
-func TestIntegrationConnPoolErrorShapes(t *testing.T) {
-	integration.CheckSkip(t)
-
-	ctr, err := testcontainers.Run(t.Context(), "mongo:latest",
-		testcontainers.WithExposedPorts("27017/tcp"),
-		testcontainers.WithEnv(map[string]string{
-			"MONGO_INITDB_ROOT_USERNAME": "mongoadmin",
-			"MONGO_INITDB_ROOT_PASSWORD": "secret",
-		}),
-		testcontainers.WithWaitStrategy(
-			wait.ForListeningPort("27017/tcp").WithStartupTimeout(time.Minute),
-		),
-	)
-	testcontainers.CleanupContainer(t, ctr)
-	require.NoError(t, err)
-
-	mp, err := ctr.MappedPort(t.Context(), "27017/tcp")
-	require.NoError(t, err)
-	port := mp.Port()
-
-	connect := func(t *testing.T, uri string, opt func(*options.ClientOptions)) *mongo.Client {
-		t.Helper()
-		o := options.Client().
-			SetConnectTimeout(5 * time.Second).
-			SetTimeout(10 * time.Second).
-			SetServerSelectionTimeout(10 * time.Second).
-			ApplyURI(uri)
-		if opt != nil {
-			opt(o)
-		}
-		client, err := mongo.Connect(o)
-		require.NoError(t, err)
-		t.Cleanup(func() { _ = client.Disconnect(context.Background()) })
-		return client
-	}
-
-	// Wait for the server to accept authenticated connections before probing
-	// error shapes, otherwise a not-yet-ready server produces the wrong error.
-	good := connect(t, "mongodb://localhost:"+port, func(o *options.ClientOptions) {
-		o.SetAuth(options.Credential{Username: "mongoadmin", Password: "secret"})
-	})
-	require.Eventually(t, func() bool {
-		return good.Ping(t.Context(), nil) == nil
-	}, time.Minute, time.Second)
-
-	t.Run("handshake_auth_failure", func(t *testing.T) {
-		bad := connect(t, "mongodb://localhost:"+port, func(o *options.ClientOptions) {
-			o.SetAuth(options.Credential{Username: "mongoadmin", Password: "wrong-password"})
-		})
-		_, err := bad.Database("TestDB").Collection("handshake").InsertOne(t.Context(), bson.M{"a": 1})
-		require.Error(t, err)
-		t.Logf("handshake failure error: %v", err)
-		// Names the exact substring isConnPoolError depends on, so a driver
-		// rewording points at this line rather than at the classifier.
-		require.Contains(t, err.Error(), "error occurred during connection handshake")
-		require.True(t, isConnPoolError(err), "real handshake failure must be classified as a pool error: %v", err)
-	})
-
-	t.Run("duplicate_key_is_not_a_pool_error", func(t *testing.T) {
-		coll := good.Database("TestDB").Collection("dupkey")
-		_, err := coll.InsertOne(t.Context(), bson.M{"_id": "dup"})
-		require.NoError(t, err)
-		_, err = coll.InsertOne(t.Context(), bson.M{"_id": "dup"})
-		require.Error(t, err)
-		t.Logf("duplicate key error: %v", err)
-		require.True(t, mongo.IsDuplicateKeyError(err), "expected a duplicate key error, got: %v", err)
-		require.False(t, isConnPoolError(err), "an ordinary write error must not trigger a reconnect: %v", err)
-	})
-
-	t.Run("server_selection_failure", func(t *testing.T) {
-		// Port 1 is never a mongod, so every operation fails in server
-		// selection rather than reaching a server.
-		dead := connect(t, "mongodb://localhost:1", func(o *options.ClientOptions) {
-			o.SetServerSelectionTimeout(2 * time.Second).SetTimeout(5 * time.Second)
-		})
-		_, err := dead.Database("TestDB").Collection("dead").InsertOne(t.Context(), bson.M{"a": 1})
-		require.Error(t, err)
-		t.Logf("server selection error: %v", err)
-		require.Contains(t, err.Error(), "server selection error")
-		require.True(t, isConnPoolError(err), "an unselectable server must be classified as a pool error: %v", err)
-	})
 }
 
 func parseInputConf(t *testing.T, conf string) *service.ParsedConfig {
